+++ piprot/test/test_piprot.py
@@ -10,8 +10,9 @@ class TestRequirementsParser(unittest.TestCase):
 
     def test_requirement_exact(self):
         with self.assertRaises(SystemExit):
-            main([open('piprot/test/files/pytz_req.txt')])
+            with open("piprot/test/files/pytz_req.txt") as f:
+                main([f])
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()